% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.r
\name{bridge_by_community}
\alias{bridge_by_community}
\title{Bridge centrality by target community}
\usage{
bridge_by_community(graph, community)
}
\arguments{
\item{graph}{p x p adjacency matrix (can be directed or undirected)}

\item{community}{character vector of length p in the SAME order as
rows/cols of graph e.g., rep(c("dep","anx","mech"), each = 3)}
}
\description{
Computes bridge strength by community pairs
}
