% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_panelGVAR.R
\name{gen_panelGVAR}
\alias{gen_panelGVAR}
\title{Generate a (multi-group) panelGVAR model}
\usage{
gen_panelGVAR(n_node = 6, n_time = 4, p_rewire = 0.5, n_group = 1)
}
\arguments{
\item{n_node}{an integer denoting the number of nodes}

\item{n_time}{an integer denoting the number of waves}

\item{p_rewire}{a numeric value between 0-1 denoting the extent of group difference}

\item{n_group}{an integer denoting the number of groups}

\item{mean_trend}{a numeric value denoting the extent of mean trends in data}

\item{p_rewire_temp}{a numeric value between 0-1 denoting the probability rewiring temporal networks across groups}

\item{p_rewire_cont}{a numeric value between 0-1 denoting the probability rewiring temporal networks across groups}
}
\value{
A list of beta, PDC, kappa and contemporaneous networks
}
\description{
This function generates a (multi-group) panel GVAR model. Currently it's focused on generating
}
\details{
\code{beta} can be transposed to obtain the temporal network;
\code{PDC} is the partial directed correlation matrix, which is a standardized version of temporal network;
\code{kappa} is the precision matrix denoting conditional (in)dependence,
which is a inverse of covariance matrix denoting the (dependence) among variables;
kappa can be further standardized to the contemporaneous networks (\code{omega_zeta_within})
}
