% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVPP-package.R
\docType{package}
\name{IVPP-package}
\alias{IVPP}
\alias{IVPP-package}
\title{IVPP: Invariance Partial Pruning Test}
\description{
An implementation of the Invariance Partial Pruning (IVPP) approach described in Du, X., Johnson, S. U., Epskamp, S. (2025) The Invariance Partial Pruning Approach to The Network Comparison in Longitudinal Data. IVPP is a two-step method that first test for global network structural difference with invariance test and then inspect specific edge difference with partial pruning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xinkaidupsy/IVPP}
  \item Report bugs at \url{https://github.com/xinkaidupsy/IVPP/issues}
}

}
\author{
\strong{Maintainer}: Xinkai Du \email{xinkai.du.xd@gmail.com} (\href{https://orcid.org/0000-0002-4158-7878}{ORCID}) [copyright holder]

}
\keyword{internal}
