% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVPP.R
\name{IVPP_panel}
\alias{IVPP_panel}
\title{The omnibus test of the invariance partial pruning (IVPP) algorithm for multi-group panel GVAR models}
\usage{
IVPP_panel(
  data,
  vars,
  idvar,
  beepvar,
  groups,
  test = c("both", "temporal", "contemporaneous"),
  net_type = c("saturated", "sparse"),
  partial_prune,
  prune_alpha = 0.01,
  p_prune_alpha = 0.01,
  estimator = "FIML",
  standardize = c("none", "z", "quantile"),
  ...
)
}
\arguments{
\item{data}{A data frame containing the long-formated panel data}

\item{vars}{A character vector of variable names}

\item{idvar}{A character string specifying subject IDs}

\item{beepvar}{A character string specifying the name of wave (time) variable}

\item{groups}{A character string specifying the name of group variable}

\item{test}{A character vector specifying the network you want to test group-equality on in the omnibus test.
Specify "both" if you want to test on both temporal or contemporaneous networks.
Specify "temporal" if you want to test only on the temporal network.
Specify "contemporaneous" if you want to test only on the contemporaneous network.
See \link{Details} for more information.}

\item{net_type}{A character vector specifying the type of networks to be compared.
Specify "saturated" if you want to estimate and compare the saturated networks.
Specify "sparse" if you want to estimate and compare the pruned networks.}

\item{partial_prune}{A logical value specifying whether to conduct partial pruning test or not.}

\item{prune_alpha}{A numeric value specifying the alpha level for the pruning (if net_type = "sparse").}

\item{p_prune_alpha}{A numeric value specifying the alpha level for the partial pruning (if partial_prune = TRUE).}

\item{estimator}{A character string specifying the estimator to be used. Must be "FIML"}

\item{standardize}{A character string specifying the type of standardization to be used.
"none" (default) for no standardization, "z" for z-scores,
and "quantile" for a non-parametric transformation to the quantiles of the marginal standard normal distribution.}

\item{...}{Additional arguments to be passed to the \code{\link[psychonetrics]{dlvm1}} function}
}
\value{
A list containing the results of IVPP and networks of all groups.
}
\description{
This function implements the IVPP algorithm to compare networks in the multi-group panel GVAR models.
The IVPP algorithm is a two-step procedure that first conducts an omnibus test of network difference and then performs partial pruning for the specific edge-level differences.
Currently supports the comparison of temporal and contemporaneous networks.
}
\details{
The comparison between the fully unconstrained (free) model and tempEq model is a test for group equality in temporal networks.
The comparison between fully constrained model (bothEq) and tempEq is a test for group equality in contemporaneous networks.
Similarly, the comparison between the free model and contEq model is a test for group equality in contemporaneous networks,
and the comparison between bothEq and contEq is a test for group equality in temporal networks.
}
