% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVPP.R
\name{panel_pp}
\alias{panel_pp}
\title{The partial_pruning test of the invariance partial pruning (IVPP) algorithm for multi-group panel GVAR models}
\usage{
panel_pp(
  data,
  vars,
  idvar,
  beepvar,
  groups,
  prune_net = c("both", "temporal", "contemporaneous"),
  prune_alpha = 0.01,
  estimator = "FIML",
  standardize = c("none", "z", "quantile"),
  ...
)
}
\arguments{
\item{data}{A data frame containing the long-formated panel data}

\item{vars}{A character vector of variable names}

\item{idvar}{A character string specifying subject IDs}

\item{beepvar}{A character string specifying the name of wave (time) variable}

\item{groups}{A character string specifying the name of group variable}

\item{prune_net}{A character vector specifying the network you want to partial prune on.
Specify "both" if you want to partial prune both temporal or contemporaneous networks.
Specify "temporal" if you want to partial prune the temporal network.
Specify "contemporaneous" if you want to partial prune the contemporaneous network.}

\item{estimator}{A character string specifying the estimator to be used. Must be "FIML"}

\item{standardize}{A character string specifying the type of standardization to be used.
"none" (default) for no standardization, "z" for z-scores,
and "quantile" for a non-parametric transformation to the quantiles of the marginal standard normal distribution.}

\item{...}{Additional arguments to be passed to the \code{\link[psychonetrics]{dlvm1}} function}
}
\value{
The results of comparison between models of different levels of constraints
}
\description{
This function implements the IVPP algorithm to compare networks in the multi-group panel GVAR models.
The IVPP algorithm is a two-step procedure that first conducts an omnibus test of network difference and then performs partial pruning for the specific edge-level differences.
This function returns the results of the omnibus test.
Currently only supports the comparison of temporal and contemporaneous networks.
}
\details{
The comparison between the fully unconstrained (free) model and tempEq model is a test for group equality in temporal networks.
The comparison between fully constrained model (bothEq) and tempEq is a test for group equality in contemporaneous networks.
Similarly, the comparison between the free model and contEq model is a test for group equality in contemporaneous networks,
and the comparison between bothEq and contEq is a test for group equality in temporal networks.
}
