% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelGVAR_gen.R
\name{panelGVAR_gen}
\alias{panelGVAR_gen}
\title{Generate Panel GVAR Dataset}
\usage{
panelGVAR_gen(
  n_node = 6,
  n_person = 500,
  n_time = 4,
  p_rewire = 0.5,
  n_group = 1,
  mean_trend = 0,
  p_rewire_temp = 0,
  p_rewire_cont = 0
)
}
\arguments{
\item{n_node}{number of nodes}

\item{n_person}{sample size}

\item{n_time}{number of waves}

\item{p_rewire}{extent of group difference}

\item{n_group}{number of groups}

\item{mean_trend}{mean trends in data}

\item{p_rewire_temp}{probability rewiring temporal networks across groups}

\item{p_rewire_cont}{probability rewiring temporal networks across groups}
}
\value{
A dataframe / datatable that has n_node variables, with subject id, wave number, and group membership specified (if multiple groups)
}
\description{
This function generates a panel GVAR dataset for ...
}
