% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_panelGVAR.R
\name{rewire}
\alias{rewire}
\title{Rewire Networks}
\usage{
rewire(net, p = 0, directed)
}
\arguments{
\item{net}{the network matrix that you want to rewire}

\item{p}{probability of rewiring}

\item{directed}{logicals. Write TRUE if the input net is a directed network and FALSE otherwise}
}
\value{
A network rewired from the input
}
\description{
This function takes in a network and rewire it with probability p
}
