% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_panelGVAR.R
\name{sim_panelGVAR}
\alias{sim_panelGVAR}
\title{Simulate data for a (multi-group) panelGVAR model}
\usage{
sim_panelGVAR(
  temp_base_ls,
  beta_base_ls,
  PDC_base_ls,
  cont_base_ls,
  n_node,
  n_person = 500,
  n_time = 3,
  n_group,
  mean_trend = 0,
  p_rewire_temp = 0,
  p_rewire_cont = 0,
  save_nets = FALSE
)
}
\arguments{
\item{n_node}{number of nodes}

\item{n_person}{an integer denoting the sample size of each group, default to 500}

\item{n_time}{number of waves}

\item{n_group}{number of groups}
}
\value{
A list of temporal and contemporaneous networks
%>%
}
\description{
This function generates data for the input (multi-group) panelGVAR model
}
